requirements.txt
accessible-pygments==0.0.5
alabaster==1.0.0
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
boto3==1.39.17
botocore==1.39.17
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docutils==0.21.2
drf-spectacular==0.28.0
furo==2025.7.19
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
oauthlib==3.3.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
roman-numerals-py==3.1.0
rpds-py==0.25.1
s3transfer==0.13.1
six==1.17.0
snowballstemmer==3.0.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.7
Sphinx==8.2.3
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
